did some todo's added obstacle states per user

+++ src/main/java/database_objects/AccountsTableRow.java
@@ -5,5 +5,5 @@ public class AccountsTableRow {
     public String username;
     public int bank;
     public String account_number;
-    public float balance;
+    public double balance;
 }
